+++ api/Models/Planet.cs
@@ -13,13 +13,13 @@ namespace api.Models
 
         public string Color { get; set; } = string.Empty;
 
-        public int XRadius { get; set; }
-        public int ZRadius { get; set; }
-        public int Size { get; set; }
+        public decimal XRadius { get; set; }
+        public decimal ZRadius { get; set; }
+        public decimal Size { get; set; }
 
-        public int Speed { get; set; }
-        public int Offset { get; set; }
-        public int RotationSpeed { get; set; }
+        public decimal Speed { get; set; }
+        public decimal Offset { get; set; }
+        public decimal RotationSpeed { get; set; }
 
         public List<TribuneArticle> TribuneArticle { get; set; } = new List<TribuneArticle>();
